+++ examples/psc/include/psc/matrix.hpp
@@ -35,7 +35,7 @@ class matrix {
 };
 
 template <typename T>
-bulk::coarray<int> lu(matrix<T>& mat) {
+bulk::coarray<size_t> lu(matrix<T>& mat) {
     auto& world = mat.world();
     auto& psi = mat.partitioning();
     auto n = mat.rows();
@@ -48,18 +48,18 @@ bulk::coarray<int> lu(matrix<T>& mat) {
     auto row_swap_k = bulk::coarray<T>(world, psi.local_size(1, t));
     auto row_swap_r = bulk::coarray<T>(world, psi.local_size(1, t));
 
-    auto pi = bulk::coarray<int>(world, psi.local_size(0, s));
+    auto pi = bulk::coarray<size_t>(world, psi.local_size(0, s));
     // partitioned according to psi_0
-    for (auto i = 0; i < psi.local_size(0, s); ++i) {
+    for (auto i = 0ul; i < psi.local_size(0, s); ++i) {
         pi[i] = psi.global(0, s, i);
     }
 
     // identify the first {l/q} such that global({l/q}) > k
-    auto ls = std::vector<int>(n);
-    auto qs = std::vector<int>(n);
-    int ridx = 0;
-    int cidx = 0;
-    for (auto k = 0; k < n; ++k) {
+    auto ls = std::vector<size_t>(n);
+    auto qs = std::vector<size_t>(n);
+    auto ridx = 0ul;
+    auto cidx = 0ul;
+    for (auto k = 0ul; k < n; ++k) {
         while (psi.global(0, s, ridx) <= k) {
             ridx++;
         }
@@ -70,16 +70,16 @@ bulk::coarray<int> lu(matrix<T>& mat) {
         qs[k] = cidx;
     }
 
-    auto pivot_rows = bulk::coarray(world, psi.grid()[0], -1);
-    auto pivot_values = bulk::coarray(world, psi.grid()[0], (T)-1);
-    auto r = bulk::var<int>(world, -1);
+    auto pivot_rows = bulk::coarray<size_t>(world, psi.grid()[0]);
+    auto pivot_values = bulk::coarray<T>(world, psi.grid()[0]);
+    auto r = bulk::var<size_t>(world);
 
-    for (auto k = 0; k < n; ++k) {
-        if (psi.owner(1, k) == t) {
+    for (auto k = 0ul; k < n; ++k) {
+        if (static_cast<size_t>(psi.owner(1, k)) == t) {
             // (0) Find maximum local element in column k
             auto r_s = -1;
             auto local_max = std::numeric_limits<T>::min();
-            for (auto j = 0; j < psi.local_size(0, s); ++j) {
+            for (auto j = 0ul; j < psi.local_size(0, s); ++j) {
                 if (std::abs(mat.at({j, psi.local(1, k)})) > local_max) {
                     local_max = std::abs(mat.at({j, psi.local(1, k)}));
                     r_s = psi.global(0, s, j);
@@ -87,7 +87,7 @@ bulk::coarray<int> lu(matrix<T>& mat) {
             }
 
             // (1) Communicate maximum element with local processor column
-            for (auto u = 0; u < psi.grid()[0]; ++u) {
+            for (auto u = 0ul; u < psi.grid()[0]; ++u) {
                 pivot_rows(psi.rank({u, t}))[s] = r_s;
                 pivot_values(psi.rank({u, t}))[s] = local_max;
             }
@@ -100,7 +100,7 @@ bulk::coarray<int> lu(matrix<T>& mat) {
                           std::max_element(pivot_values.begin(), pivot_values.end()));
 
             // (3) Communicate global maximum with local processor row
-            for (auto v = 0; v < psi.grid()[1]; ++v) {
+            for (auto v = 0ul; v < psi.grid()[1]; ++v) {
                 r(psi.rank({s, v})) = pivot_rows[best_index];
             }
         } else {
@@ -109,11 +109,13 @@ bulk::coarray<int> lu(matrix<T>& mat) {
         world.sync();
 
         // (4) Communicate permutation update
-        if (psi.owner(0, k) == s) {
-            pi(psi.rank({psi.owner(0, r), t}))[psi.local(0, r)] = pi[psi.local(0, k)];
+        if (static_cast<size_t>(psi.owner(0, k)) == s) {
+            pi(psi.rank({static_cast<size_t>(psi.owner(0, r)), t}))[psi.local(0, r)] =
+            pi[psi.local(0, k)];
         }
-        if (psi.owner(0, r) == s) {
-            pi(psi.rank({psi.owner(0, k), t}))[psi.local(0, k)] = pi[psi.local(0, r)];
+        if (static_cast<size_t>(psi.owner(0, r)) == s) {
+            pi(psi.rank({static_cast<size_t>(psi.owner(0, k)), t}))[psi.local(0, k)] =
+            pi[psi.local(0, r)];
         }
         world.sync();
 
@@ -121,15 +123,15 @@ bulk::coarray<int> lu(matrix<T>& mat) {
         // ... not necessary
 
         // (6) Communicate row k and r
-        if (psi.owner(0, k) == s) {
-            for (auto j = 0; j < psi.local_size(1, t); ++j) {
-                row_swap_k(psi.rank({psi.owner(0, r), t}))[j] =
+        if (static_cast<size_t>(psi.owner(0, k)) == s) {
+            for (auto j = 0ul; j < psi.local_size(1, t); ++j) {
+                row_swap_k(psi.rank({static_cast<size_t>(psi.owner(0, r)), t}))[j] =
                 mat.at({psi.local(0, k), j});
             }
         }
-        if (psi.owner(0, r) == s) {
-            for (auto j = 0; j < psi.local_size(1, t); ++j) {
-                row_swap_r(psi.rank({psi.owner(0, k), t}))[j] =
+        if (static_cast<size_t>(psi.owner(0, r)) == s) {
+            for (auto j = 0ul; j < psi.local_size(1, t); ++j) {
+                row_swap_r(psi.rank({static_cast<size_t>(psi.owner(0, k)), t}))[j] =
                 mat.at({psi.local(0, r), j});
             }
         }
@@ -137,26 +139,26 @@ bulk::coarray<int> lu(matrix<T>& mat) {
         world.sync();
 
         // (7) Locally swap row k and r
-        if (psi.owner(0, k) == s) {
-            for (auto j = 0; j < psi.local_size(1, t); ++j) {
+        if (static_cast<size_t>(psi.owner(0, k)) == s) {
+            for (auto j = 0ul; j < psi.local_size(1, t); ++j) {
                 mat.at({psi.local(0, k), j}) = row_swap_r[j];
             }
         }
-        if (psi.owner(0, r) == s) {
-            for (auto j = 0; j < psi.local_size(1, t); ++j) {
+        if (static_cast<size_t>(psi.owner(0, r)) == s) {
+            for (auto j = 0ul; j < psi.local_size(1, t); ++j) {
                 mat.at({psi.local(0, r), j}) = row_swap_k[j];
             }
         }
         // (8) Communicate pivot
         if (psi.owner({k, k}) == world.rank()) {
-            for (auto u = 0; u < psi.grid()[0]; ++u) {
+            for (auto u = 0ul; u < psi.grid()[0]; ++u) {
                 pivot(psi.rank({u, t})) = mat.at(psi.local({k, k}));
             }
         }
         world.sync();
 
         // (9) Scale column
-        if (psi.owner(1, k) == t) {
+        if (static_cast<size_t>(psi.owner(1, k)) == t) {
             // local row index, local col size, loop
             for (auto i = ls[k]; i < psi.local_size(0, s); ++i) {
                 mat.at({i, psi.local(1, k)}) /= pivot;
@@ -164,10 +166,10 @@ bulk::coarray<int> lu(matrix<T>& mat) {
         }
 
         // (10) Horizontal, communicate column k
-        if (psi.owner(1, k) == t) {
-            for (auto v = 0; v < psi.grid()[1]; ++v) {
+        if (static_cast<size_t>(psi.owner(1, k)) == t) {
+            for (auto v = 0ul; v < psi.grid()[1]; ++v) {
                 auto target_rank = psi.rank({s, v});
-                for (auto i = 0; i < psi.local_size(0, s); ++i) {
+                for (auto i = 0ul; i < psi.local_size(0, s); ++i) {
                     col_k(target_rank)[i] = mat.at({i, psi.local(1, k)});
                 }
             }
@@ -175,10 +177,10 @@ bulk::coarray<int> lu(matrix<T>& mat) {
         world.sync();
 
         // (10b) Vertical, communicate row k
-        if (psi.owner(0, k) == s) {
-            for (auto u = 0; u < psi.grid()[0]; ++u) {
+        if (static_cast<size_t>(psi.owner(0, k)) == s) {
+            for (auto u = 0ul; u < psi.grid()[0]; ++u) {
                 auto target_rank = psi.rank({u, t});
-                for (auto j = 0; j < psi.local_size(1, t); ++j) {
+                for (auto j = 0ul; j < psi.local_size(1, t); ++j) {
                     row_k(target_rank)[j] = mat.at({psi.local(0, k), j});
                 }
             }
@@ -209,8 +211,8 @@ void spy(matrix<T>& mat) {
     auto x = phi.local_size(world.rank())[0];
     auto y = phi.local_size(world.rank())[1];
 
-    for (auto i = 0; i < x; ++i) {
-        for (auto j = 0; j < y; ++j) {
+    for (auto i = 0ul; i < x; ++i) {
+        for (auto j = 0ul; j < y; ++j) {
             xs(0)[bulk::util::flatten<2>(phi.global_size(), phi.global({i, j}, {s, t}))] =
             mat.at({i, j});
         }
@@ -219,9 +221,9 @@ void spy(matrix<T>& mat) {
     world.sync();
 
     if (world.rank() == 0) {
-        for (auto i = 0; i < mat.rows(); ++i) {
+        for (auto i = 0ul; i < mat.rows(); ++i) {
             std::cout << "[ " << std::fixed << std::setprecision(2);
-            for (auto j = 0; j < mat.cols(); ++j) {
+            for (auto j = 0ul; j < mat.cols(); ++j) {
                 std::cout << xs[bulk::util::flatten<2>(phi.global_size(), {i, j})] << " ";
             }
             std::cout << "]\n";