+++ include/bulk/partitionings/tree.hpp
@@ -30,7 +30,8 @@ class tree_partitioning : public rectangular_partitioning<D, 1> {
      * `data_size`: the global number of processors along each axis
      */
     tree_partitioning(index_type<D> data_size, int procs, util::binary_tree<util::split>&& splits)
-    : rectangular_partitioning<D, 1>(data_size, {procs}), splits_(std::move(splits)) {
+    : rectangular_partitioning<D, 1>(data_size, {static_cast<size_t>(procs)}),
+      splits_(std::move(splits)) {
         // resize extents and origins
         origins_.resize(procs);
         extents_.resize(procs);
@@ -78,8 +79,8 @@ class tree_partitioning : public rectangular_partitioning<D, 1> {
         // we encode the path to the final volume as a bit pattern, that will be
         // the processor id
         auto node = splits_.root.get();
-        int proc = 0;
-        int depth = 0;
+        size_t proc = 0;
+        size_t depth = 0;
         while (node) {
             if (xs[node->value.d] <= node->value.a) {
                 node = node->left.get();