+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - YYYY-MM-DD
+
+### Breaking changes
+
+- `bulk::index` now represents indices as `size_t`
+
 ## [2.0.0] - 2020-03-27
 
 ### Breaking changes